Move for_tests module into its own file
